Add virtual NUMA support
* Added NUMA topology and processor related HCS schema definitions
* New annotations "io.microsoft.virtualmachine.computetopology.numa.default"
and "io.microsoft.virtualmachine.computetopology.numa.settings" to be
able to specify the virtual NUMA topology.
* Validate virtual NUMA topology passed in for obvious flaws.
* New UVM method hostProcessorInfo that will call HcsGetServiceProperties
and return the hosts processor topology as well as the total number of
logical processors on the system. This is to fix the bug where we were
incorrectly capping a users requested CPU count on > 1 NUMA node systems
as runtime.NumCPU only sees the LPs from the processor group it is run on.
* For process isolated containers instead of using runtime.NumCPU to
normalize the requested processor count, use the win32 API function
GetMaximumProcessorCount as it will see all the LPs in all processor groups.
This is to account for minroot setups where the host is only able to see a
certain amount of LPs and the rest are reserved for guests.
* Refactored checking for validations/unsupported configurations when creating
the UVM doc.
* Run 'Go mod vendor' to get NUMA schema to use in tests.
* Added tests for NUMA validation.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>