Move HCS resource path definitions to their own package
As we need to make HCS calls in the new /internal/vm/hcs package we need access
to the virtual machine resource paths. They're currently unexported, but it makes
more sense to have all of the resource paths in one place so that's what I've done.

* Move all of the virtual machine and silo containe resource paths to their own package
and update the uses of them everywhere.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>