Move opengcs/internal to /internal
The debug package is useful and not strictly related to guest behavior so I've moved this to
a new debug package in internal, however for everything that's either Linux specific or opengcs
specific I've created a new `guest` package that houses these. This includes the kmsg package
(linux specific) the hcs v2 runtime guest code, all of the storage functionality that ends up
leading to mount syscalls, and guest side vmbus/pci related code.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>