Build gcs and gcstools binaries out of cmd
This commit moves the code for the main opengcs binary and the gcstools binary
to cmd so we can build the binaries the same way we do for the all of the other
binaries we care about.

The Makefile will be edited in a future commit to be aware of this rearrange.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>